Interface with Library_Beam to perform realistic calculations using the rotation matrix calculated by interpolated angular displacements
Blade axis specification in couplingProperties is in the OpenFOAM frame;
bug fixed for deflection in the wrong direction.

New quantities needed to be saved are the position vector along the beam
axis (x1) and the surface node offset vector (p)

Saved information goes into either the project directory or the
processor directories (in parallel).